requirements.txt
nltk
tabulate
numpy
pandas
notebook
torch
openai
scipy
jsonlines
python-dotenv
# Hugging Face
datasets
evaluate
accelerate==0.21.0 
peft==0.4.0 
bitsandbytes==0.40.2 
transformers==4.31.0 
trl==0.4.7